+++ tests/fixtures/appToken.js
@@ -1,4 +1,0 @@
-module.exports = {
-  APP_TOKEN_VALID: process.env.APP_TOKEN,
-  APP_TOKEN_INVALID: 'APP_TOKEN_INVALID'
-};