+++ tests/unit/expectations/request.js
@@ -1,0 +1,10 @@
+const expectRequestHeaders = (setHeadersFn, appToken) => {
+  expect(setHeadersFn).toHaveBeenCalledWith({
+    'Content-Type' : 'application/json',
+    'Authorization': 'Bearer ' + appToken
+  });
+}
+
+module.exports = {
+  expectRequestHeaders
+};